requirements.txt
Flask>=0.11.0
requests
paho-mqtt